+++ test/tests/filterFuzzTest.pl
@@ -24,7 +24,7 @@ while (<$fh>) {
 }
 close $fh;
 
-push @topics, "nosuchtopic";
+push @$topics, "nosuchtopic";
 
 sub genRandomFilterGroup {
     my $useLimit = shift;
@@ -136,7 +136,7 @@ sub testScan {
 
     my $headCmd = @$fg == 1 && $fg->[0]->{limit} ? "| head -n $fg->[0]->{limit}" : "";
 
-    my $resA = `./strfry export --reverse 2>/dev/null | perl test/dumbFilter.pl '$fge' $headCmd | jq -r .id | sort | sha256sum`;
+    my $resA = `./strfry export --reverse 2>/dev/null | perl test/utils/dumbFilter.pl '$fge' $headCmd | jq -r .id | sort | sha256sum`;
     my $resB = `./strfry scan --pause 1 --metrics '$fge' | jq -r .id | sort | sha256sum`;
 
     print "$resA\n$resB\n";
